+++ src/main/java/com/cibertec/repository/LibroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.cibertec.model.Libro;
 @Repository
-public interface LibroRepository extends JpaRepository<Libro, Long> {
+public interface LibroRepository extends JpaRepository<Libro, Integer> {
     // Aquí puedes crear métodos personalizados, como buscar por título
 }
